Add core chess game abstract classes
These classes are the core of Chess Plus Plus. The goal is to make it so that these base classes encompass a majority of what is needed for the game to run. The core game manager will assume that any sort of board, game state, cell, etc. that is passed to it inherits from these core classes. That way, if a new gamemode is desired in the future, the only thing to do is to inherit from one of these classes and register it with the system.